Changed the name and reformatted Costa's test
